* Fixed MPI compilation issue. * Measuring time intervals requires double precision.

+++ src/basic/global.F90
@@ -125,7 +125,7 @@ module global_m
 
   ! The stack.
   character(len=40), public          :: sub_stack(50)
-  FLOAT, public                      :: time_stack(50)
+  real(8), public                    :: time_stack(50)
   integer, public                    :: no_sub_stack = 0
 
   ! should we run in debug mode
